requirements.txt
annotated-types==0.7.0
anyio==4.7.0
click==8.1.8
fastapi==0.115.6
h11==0.14.0
idna==3.10
numpy==2.2.1
pandas==2.2.3
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
